requirements.txt
aiohttp==3.8.3; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.0"
appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.7"
async-timeout==4.0.2; python_version >= "3.6" and python_full_version >= "3.6.0"
attrs==22.1.0; python_version >= "3.6" and python_full_version >= "3.6.0"
backcall==0.2.0; python_version >= "3.7"
backoff==1.11.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
certifi==2022.9.14; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
click-help-colors==0.8.1
click-option-group==0.5.3; python_version >= "3.6" and python_version < "4"
click-spinner==0.1.10
click==7.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
comma==0.5.4; python_version >= "3.6" and python_version < "4.0"
decorator==5.1.1; python_version >= "3.7"
frozenlist==1.3.1; python_version >= "3.7" and python_full_version >= "3.6.0"
idna==3.4; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
ipython==7.34.0; python_version >= "3.7"
jedi==0.18.1; python_version >= "3.7"
jinja2==2.11.3; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
loguru==0.5.3; python_version >= "3.5"
markupsafe==2.0.1; python_version >= "3.6"
matplotlib-inline==0.1.6; python_version >= "3.7"
multidict==6.0.2; python_version >= "3.7" and python_full_version >= "3.6.0"
parso==0.8.3; python_version >= "3.7"
pbr==5.10.0; python_version >= "3.6"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
pillow==9.0.1; python_version >= "3.7"
ply==3.11; python_version >= "3.6"
prompt-toolkit==3.0.31; python_full_version >= "3.6.2" and python_version >= "3.7"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7"
pygments==2.13.0; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-dotenv==0.14.0
pyyaml==5.4.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
requests==2.28.1; python_version >= "3.7" and python_version < "4"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
slack-scim==1.1.0; python_version >= "3.6"
slackclient==2.9.4; python_full_version >= "3.6.0"
tqdm==4.64.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
traitlets==5.4.0; python_version >= "3.7"
urllib3==1.26.12; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0" and python_version < "4")
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.7"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
yaql==2.0.0; python_version >= "3.6"
yarl==1.8.1; python_version >= "3.7" and python_full_version >= "3.6.0"
